Multi Games w/o closing app(RESET), multi single P games not working yet
When the reset or Main Menu buttons are clicked, the scene coordinator is re instantiated to simplify reset procedures. When playing multiple single player games, there is an error that has to do with the response thread in game controller being called upon by different Scene coordinators, figure out a way to close the thread after the game is reset.